README.md
Day 7 of 100 Days of Code. Here we have a pretty straightforward game of Hangman. There is a separate file containing a big list of words, acting as the name bank, which was imported.

It tells you what letter you have guessed and if it is in the word or not. If the letter was previously already guessed (correctly), it lets you know. If it previously guessed incorrectly, not only does it let you know, but it counts as a point against you, yet again. A user can scroll up to see the history of the game to try to avoid that. And if you lose, it tells you what the word was.